fix: correct publicpath in build
