+++ .changeset/angry-owls-retire.md
@@ -1,0 +1,5 @@
+---
+'@module-federation/sdk': patch
+---
+
+feat(sdk): add ssrPublicPath field
