+++ .changeset/early-toys-occur.md
@@ -1,0 +1,5 @@
+---
+'@module-federation/modern-js': patch
+---
+
+fix(modern-js-plugin): correct publicpath in build
