+++ packages/modernjs/src/constant.ts
@@ -1,3 +1,2 @@
-export const MODERN_JS_SERVER_DIR = 'bundles';
 export const LOCALHOST = 'localhost';
 export const PLUGIN_IDENTIFIER = '[ Modern.js Module Federation ]';
